Adds a regex filtering system to "list" site types.
This is an optional regex value that one may add in configuration
file for sites of type "list" (see version.yaml provided as an
example). Two parts are to be matched in this regex : the first
one is assumed to be project name and the second one is assumed
to be the version.