Refactoring to get simpler code
Renamed project_entry to site_entry to better reflect the content of
that variable: it contains the value of 'entry field for the whole
site).

Extracted the code that decides if last_checked is used in either
site or project definition and make an new function called
is_one_entry_field_value_egal_to_last_check() with it.